+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { RouterProvider } from "react-router-dom";
-import router from "./routes/router";
+import router from "./routes/routers";
 import { Toaster } from "react-hot-toast";
 
 function App() {
